requirements.txt
flask
twilio
flask-socketio
python-dotenv
watchdog